tune data loading; always record stats regardless of Exposed
There is a balance for recording stats.  On the one hand, if you know
500 words, forcing you to use Journey Mode to expose them before other
activities is problematic.  On the other hand, if you don't know any
words, and try (and fail / guess correctly) on other activities, it
could mess up the Journey Mode progress.

On balance, we do need the stats for everything.  Perhaps a new Journey
Mode algorithm will add additional data if this causes practical
problems.